Change set up information regarding majority board
Fix interaction between module 6 & 7. Determine which side of the majority board is face up.
Move module 2.III information to after module 6 and 7 to help flow of rules in the case that you have either module 6 or 7 in the game and the majority board for 2.III. 
Clear up some corner cases with module 2.III.
Update mid term scoring rules and exploration rules for 5x2 and 952.