Update to interaction between module 6 and 7
Removed some redundant lines from the resident phase. 
When playing with module 7 and with module 6 in TOP I or II you should use your road network to determine majorities. code has been altered to reflect that. (it may seem that there are combinations where module 8 would interfere with this but this can only happen with 87x games and then 6 would not be in top I or II.)
The rules for majorities with module 6.I or II combined with Module 4 are explained in the module 6 section. So the line for module 4 in the module 7 rules is unnecessary (and incorrect) with module 6.I or 6.II (ie."If the player conquers the tile of another player, he counts that tile on the majority board, while the other player subtracts it." )
logic added to stop duplication due to identical lines of rules for module 6.III with module 2 and module 6.III with module 7.1.