Update rules to reflect nov 15 Errata/FAQ
https://boardgamegeek.com/filepage/123135/2f-spielestronghold-games-504-faq-clarifications-a
lots of small changes, the change that is the most common in this document relates to module 2 and city cards. The rules now refer to population on city cards not residents, (population includes residents and settlements) as you put a settlement on the city card that matches your starting capital.
Many of the changes in the errata had already been included based on threads on bgg.