update phase 3 d name tokens
Als fixed rules conflict in lines 830/831 as suggested in bgg forums. (if playing with modules 4 and 2 with 5.II the module in pos I takes priority) This hasn't been explicitly ruled on by friese or henning but it is a general rule, when you have a rules conflict go in favour of the higher module.


Additional token "sagent"" (second agent) created for when "he" follows an owner token as the he would often be refering to the agent in the case of company games it should be "the president". let me know if this is getting out of control with these tokens.