Phase 3E reorg
the biggest change is to the placing settlements without residents rule. I have made this rule always show with a note that it may not apply. It wasn't showing up for all the combinations where it should and in several it was showing up twice. New rule mentions that it is not always necessary.