update phase 3 F - H name tokens plus some clarification.
Additional change to rules for module 2.II income (and 29x income) with module 4. This works with neutralization and conquering so I have separated this out from the other rules.
Rule for 3.I/3.II is irrelevant with module companies.
rule for 89x income has errata, you may not exchange goods in this phase.