Update rules to reflect errata 452 254
There was uncertainty over which rule was dominant in 452 and 254. but the change of wording of the rule in the errata has elimanted the rules conflict.  the rule for module 4 ensures that you remove the resident from the board but doesn't overide the number of explored hexes, so The rule for module 2 is always in effect in 25x and x52 games, even games with module 4,